fix(server): cast elapsed to i64 in WS/TCP online status check
The databk-specific WS/TCP online-status path still cast elapsed to
i32, mismatching REG_TIMEOUT which was widened to i64 for #676. Cast to
i64 to fix the CI compile error (E0308 at rendezvous_server.rs:681).

+++ src/rendezvous_server.rs
@@ -675,7 +675,7 @@ impl RendezvousServer {
                     let mut states = BytesMut::zeroed((peers.len() + 7) / 8);
                     for (i, peer_id) in peers.iter().enumerate() {
                         if let Some(peer) = self.pm.get_in_memory(peer_id).await {
-                            let elapsed = peer.read().await.last_reg_time.elapsed().as_millis() as i32;
+                            let elapsed = peer.read().await.last_reg_time.elapsed().as_millis() as i64;
                             let states_idx = i / 8;
                             let bit_idx = 7 - i % 8;
                             if elapsed < REG_TIMEOUT {
